+++ crates/px-configuration/src/lib.rs
@@ -3,4 +3,5 @@ mod profile;
 mod project;
 
 // pub use global::*;
+pub use profile::*;
 pub use project::*;