+++ CHANGELOG.md
@@ -18,6 +18,7 @@
 * Add `px project init` to create a folder and config file
 * Add `px project add` to add an observation to the project and start filling out linear stack creation
 * Add `px project stack` and basic implementation with a number of todo's
+* Add `px profile init` and a convention plus profile config file
 
 ## 0.1.3